refactor: extract pref_store.rs from brain/preferences.rs
Split preferences.rs (1,959 lines) into 2 focused modules:
- pref_store.rs (363 lines): persistence, JSON serialization, adaptive_threshold
- preferences.rs (1,644 lines): distillation algorithms, types, re-exports

All existing import paths preserved via pub use re-exports.
