+++ src/brain/mod.rs
@@ -8,6 +8,7 @@ pub mod evals;
 pub mod insights;
 pub mod mailbox;
 pub mod metrics;
+pub mod pref_store;
 pub mod preferences;
 pub mod prompts;
 pub mod retrieval;
